Renamed module package. Also tested entry points for setup.py develop and install

+++ reference_file_testing_tool/db.py
@@ -1,4 +1,4 @@
-""" Database utility scripts
+"""Database utility scripts.
 
 Usage:
   db_utils (create | remove) <db_path>
@@ -13,31 +13,22 @@ Options:
   --version     Show version.
 """
 
-from astropy.io import fits
-from docopt import docopt
 import glob
 import os
+
+from astropy.io import fits
+from docopt import docopt
 from sqlalchemy import Column, Integer, String, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
-# import logging
-
-__all__ = ['TestData', 'data_exists', 'load_session', 'create_test_data_db',
-           'add_test_data']
-
-# log = logging.getLogger(__name__)
-# log.setLevel(logging.DEBUG)
 
 Base = declarative_base()
 
-REFTEST_DATA_DB = os.environ.get('REFTEST_DB')
-
 class TestData(Base):
     __tablename__ = 'test_data'
-    # Here we define columns for the exposures table which will just contain
-    # some basic fits header information.
-    id = Column(Integer, primary_key=True)
-    filename = Column(String(250))
+
+    filename = Column(String(100), primary_key=True)
+    path = Column(String(200))
     DATE_OBS = Column(String(10))
     TIME_OBS = Column(String(12))
     INSTRUME = Column(String(20))
@@ -57,11 +48,12 @@ class TestData(Base):
 
 
     def __init__(self, filename):
-        # you don't have to create an __init__()
-        # but it makes it easier to create a new row
-        # from a FITS file
-        self.filename = os.path.abspath(filename)
+        # Populates row when instance of class is envoked.
         header = fits.getheader(filename)
+        path, name = os.path.split(filename)
+        
+        self.filename = name
+        self.path = path
         self.DATE_OBS = header.get('DATE-OBS')
         self.TIME_OBS = header.get('TIME-OBS')
         self.INSTRUME = header.get('INSTRUME')
@@ -79,6 +71,7 @@ class TestData(Base):
         self.SUBSIZE1 = header.get('SUBSIZE1')
         self.SUBSIZE2 = header.get('SUBSIZE2')
 
+
 def data_exists(fname, session):
     """
     Check if there is already a dataset with the proposed dataset's parameters
@@ -92,9 +85,9 @@ def data_exists(fname, session):
 
     Returns
     -------
-        True if there is no 
-        
+        True if there are no matches
     """
+
     header = fits.getheader(fname)
     args = {}
     args['INSTRUME'] = header.get('INSTRUME')
@@ -128,22 +121,17 @@ def load_session(db_path=None):
     Returns
     -------
     session: sqlalchemy.orm.Session
-
     """
+
     # set up a session with the database
     if db_path is None:
-        if REFTEST_DATA_DB is None:
-            print("REFTEST_DB is None, please specify a database")
-            return None
-        else:
-            db_path = REFTEST_DATA_DB
-
-    engine = create_engine('sqlite:///{}'.format(db_path), echo=False)
-    metadata = Base.metadata
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    print("Connected to DB at {}".format(db_path))
-    return session
+        print("db_path = None, SUPPLY ABSOLUTE PATH TO DB!")
+    else:
+        engine = create_engine('sqlite:///{}'.format(db_path), echo=False)
+        Session = sessionmaker(bind=engine)
+        session = Session()
+        print("Connected to DB at {}".format(db_path))
+        return session
 
 
 def create_test_data_db(db_path):
@@ -160,6 +148,7 @@ def create_test_data_db(db_path):
     Base.metadata.create_all(engine)
     print("CREATED DB: {}".format(db_path))
 
+
 def add_test_data(file_path, db_path=None, force=False, replace=False):
     """
     Add files to the test data DB.
@@ -167,9 +156,9 @@ def add_test_data(file_path, db_path=None, force=False, replace=False):
     Parameters
     ----------
     file_path: str
-        Globable file string for test data
+        Absolute path for data file to add
     db_path: str
-        Path to database on local machince
+        Absolute path to database on local machince
     force: bool
         Force add to db even if file shares same field entries
     replace: bool
@@ -184,21 +173,22 @@ def add_test_data(file_path, db_path=None, force=False, replace=False):
     for fname in glob.glob(file_path):
         query_result = data_exists(fname, session)
         if query_result.count() != 0 and not (force or replace):
-            print('There is already test data with the same parameters. To add the data anyway set force=True')
+            print("There is already test data with the same parameters. To add the data anyway use 'db_utils force <db_path> <file_path>'")
         elif query_result and replace:
             session.delete(query_result.first())
             session.add(TestData(fname))
             session.commit()
-            print('Replaced {} with {}'.format(query_result.first().filename,
+            print("Replaced {} with {}".format(query_result.first().filename,
                                                file_path))
         else:
             new_test_data = TestData(fname)
             session.add(new_test_data)
             session.commit()
-            print('Added {} to database'.format(file_path))
+            print("Added {} to database".format(file_path))
+
 
 def main():
-    """ Main to parse command line entries.
+    """Main to parse command line entries.
 
     Parameters
     ----------
@@ -208,6 +198,8 @@ def main():
     -------
     None
     """
+
+    # Get docopt arguments..
     args = docopt(__doc__, version='0.1')
 
     # Parse command line arguments
@@ -219,7 +211,10 @@ def main():
                       force=args['force'], 
                       replace=args['replace'])
     else:
-        # Make sure to check db_path ends with .db, dont want to delete other
-        # files that aren't databases....
-        os.remove(args['<db_path>'])
-        print("DELETED DB {}".format(args['<db_path>']))
+        # Maybe try to open a session with file and if it fails, don't delete
+        # Else, delete.
+        if args['<db_path>'].endswith('.db'):
+            os.remove(args['<db_path>'])
+            print("DELETED DB {}".format(args['<db_path>']))
+        else:
+            print('TO DELETE DB, FILE NEEDS TO END WITH ".db"')
